se elimina capa de dao y se reemplaza por repository jpa spring data
