+++ src/main/java/com/github/erodriguezg/springbootangular/entities/Comuna.java
@@ -1,24 +1,12 @@
-package com.github.erodriguezg.springbootangular.services.entities;
+package com.github.erodriguezg.springbootangular.entities;
 
+import javax.persistence.*;
+import javax.validation.constraints.Size;
 import java.io.Serializable;
 import java.util.Objects;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.NamedQueries;
-import javax.persistence.NamedQuery;
-import javax.persistence.Table;
-import javax.validation.constraints.Size;
-
-
 @Entity
 @Table(name = "comuna")
-@NamedQueries({
-	@NamedQuery(name = "Comuna.traerPorProvincia", query = "select c from Comuna c where c.provincia = :provincia order by c.nombre asc ")
-})
 public class Comuna implements Serializable {
 
     private static final long serialVersionUID = 6876833168788473163L;
