+++ src/main/java/com/github/erodriguezg/springbootangular/repository/PerfilUsuarioRepository.java
@@ -1,0 +1,12 @@
+package com.github.erodriguezg.springbootangular.repository;
+
+import com.github.erodriguezg.springbootangular.entities.PerfilUsuario;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface PerfilUsuarioRepository extends JpaRepository<PerfilUsuario, Integer> {
+
+
+
+}
