+++ src/main/java/com/github/erodriguezg/springbootangular/services/ComunaService.java
@@ -1,14 +1,42 @@
 package com.github.erodriguezg.springbootangular.services;
 
-import java.util.List;
-
+import com.github.erodriguezg.springbootangular.entities.Comuna;
+import com.github.erodriguezg.springbootangular.entities.Provincia;
+import com.github.erodriguezg.springbootangular.repository.ComunaRepository;
 import com.github.erodriguezg.springbootangular.services.dto.ComunaDto;
 import com.github.erodriguezg.springbootangular.services.dto.ProvinciaDto;
+import com.github.erodriguezg.springbootangular.services.mappers.ComunaDtoMapper;
+import com.github.erodriguezg.springbootangular.services.mappers.ProvinciaDtoMapper;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
-public interface ComunaService {
+import java.util.List;
+import java.util.stream.Collectors;
 
-	ComunaDto traerPorId(Integer idComuna);
+@Service
+@Transactional(readOnly = true)
+public class ComunaService {
 
-	List<ComunaDto> traerPorProvincia(ProvinciaDto provinciaDto);
+    @Autowired
+    private ComunaRepository comunaRepository;
+
+    @Autowired
+    private ComunaDtoMapper comunaMapper;
+
+    @Autowired
+    private ProvinciaDtoMapper provinciaMapper;
+
+    public ComunaDto traerPorId(Integer idComuna) {
+        Comuna comuna = comunaRepository.findById(idComuna).orElse(null);
+        return this.comunaMapper.toDto(comuna);
+    }
+
+    public List<ComunaDto> traerPorProvincia(ProvinciaDto provinciaDto) {
+        Provincia provincia = this.provinciaMapper.toEntidad(provinciaDto);
+        return comunaRepository.findByProvincia(provincia).stream()
+                .map(this.comunaMapper::toDto)
+                .collect(Collectors.toList());
+    }
 
 }
