--- src/main/java/com/github/erodriguezg/springbootangular/services/PerfilService.java
@@ -1,11 +1,0 @@
-package com.github.erodriguezg.springbootangular.services;
-
-import java.util.List;
-
-import com.github.erodriguezg.springbootangular.services.dto.PerfilDto;
-
-public interface PerfilService {
-
-	List<PerfilDto> traerTodos();
-
-}
