--- src/main/java/com/github/erodriguezg/springbootangular/services/entities/Archivo.java
@@ -1,14 +1,0 @@
-package com.github.erodriguezg.springbootangular.services.entities;
-
-public class Archivo {
-
-    private Long idArchivo;
-
-    private byte[] bytesContent;
-
-    private String fileName;
-
-    private String contentType;
-
-
-}
