--- src/main/java/com/github/erodriguezg/springbootangular/services/entities/FotoPersona.java
@@ -1,10 +1,0 @@
-package com.github.erodriguezg.springbootangular.services.entities;
-
-public class FotoPersona {
-
-    private Persona persona;
-
-    private Archivo archivo;
-
-
-}
