--- src/main/java/com/github/erodriguezg/springbootangular/services/entities/FotoPersonaPK.java
@@ -1,10 +1,0 @@
-package com.github.erodriguezg.springbootangular.services.entities;
-
-public class FotoPersonaPK {
-
-    private Long idPersona;
-
-    private Long idArchivo;
-
-
-}
