--- src/main/java/com/github/erodriguezg/springbootangular/services/entities/Region.java
@@ -1,59 +1,0 @@
-package com.github.erodriguezg.springbootangular.services.entities;
-
-import java.io.Serializable;
-import java.util.Objects;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.NamedQueries;
-import javax.persistence.NamedQuery;
-import javax.persistence.Table;
-
-@Entity
-@Table(name = "region")
-@NamedQueries({
-	@NamedQuery(name = "Region.traerTodas", query = "select r from Region r order by r.nombre asc ")
-})
-public class Region implements Serializable {
-
-	private static final long serialVersionUID = -4189202485773018556L;
-
-	@Id
-	@Column(name = "id_region")
-	private Integer id;
-
-	@Column(name = "nombre")
-	private String nombre;
-
-	public Integer getId() {
-		return id;
-	}
-
-	public void setId(Integer id) {
-		this.id = id;
-	}
-
-	public String getNombre() {
-		return nombre;
-	}
-
-	public void setNombre(String nombre) {
-		this.nombre = nombre;
-	}
-
-	@Override
-	public boolean equals(Object o) {
-		if (this == o)
-			return true;
-		if (o == null || getClass() != o.getClass())
-			return false;
-		Region region = (Region) o;
-		return Objects.equals(id, region.id);
-	}
-
-	@Override
-	public int hashCode() {
-		return Objects.hash(id);
-	}
-}
