+++ src/main/java/com/github/erodriguezg/springbootangular/services/mappers/UsuarioDtoMapper.java
@@ -1,7 +1,7 @@
 package com.github.erodriguezg.springbootangular.services.mappers;
 
 import com.github.erodriguezg.springbootangular.services.dto.UsuarioDto;
-import com.github.erodriguezg.springbootangular.services.entities.Usuario;
+import com.github.erodriguezg.springbootangular.entities.Usuario;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
